pyproject.toml
[project]
name = "smallstep-ca"
version = "0.1.0"
description = ""
authors = [
    {name = "Aaron D Borden",email = "adborden@a14n.net"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "molecule (>=26.4.0,<27.0.0)",
    "molecule-plugins[podman] (>=25.8.12,<26.0.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-testinfra (>=10.2.2,<11.0.0)"
]
